Move get_list_of_videos to common_utils.py
